fix(build): unblock Android and iOS builds for the home-screen widget (#141)
Android: home_widget 0.8.1 pins its own subproject to
kotlinOptions { jvmTarget = "1.8" } while its dynamic work-runtime-ktx /
coroutines dependencies carry JVM 11 inline bytecode, failing
:home_widget:compileDebugKotlin. The root build.gradle.kts now raises
every Kotlin plugin subproject to jvmTarget 17 (plus matching Java 17
compile options for AGP's consistency check), registered before the
evaluationDependsOn(:app) block that force-evaluates :app. The same
block raises library compileSdk to 37 when lower — home_widget's
glance-appwidget 1.3.0-alpha02 AAR metadata demands it, and 37 is
already :app's pin.

Also fixes a failure this unmasked in :app:compileDebugKotlin (never
reached while the plugin compile failed): LunarLogWidgetProvider
narrowed the Long elapsed-day count to Int once so the untilNext
arithmetic stays in the payload's Int domain (approxDaysLabel takes
Int).

iOS: project.pbxproj referenced LunarLogWidgetBundle.swift, which does
not exist on disk — LunarLogWidget.swift already declares the @main
WidgetBundle. Removed the four stale references (PBXBuildFile,
PBXFileReference, group child, Sources phase) so the target compiles
exactly the one Swift file present.

Verified: flutter build apk --debug green; flutter analyze clean;
flutter test 6271 passing; quality gate PASS.

+++ android/app/src/main/kotlin/com/wjdavis5/lunarlog/LunarLogWidgetProvider.kt
@@ -86,8 +86,11 @@ class LunarLogWidgetProvider : AppWidgetProvider() {
         val asOf = prefs.getString(KEY_AS_OF, null)?.let { parseDate(it) }
         val elapsed = if (asOf != null) daysBetween(asOf, LocalDate.now()) else 0L
         // Defensive: a device clock rollback renders the stored values
-        // unchanged rather than counting backwards.
-        val rolled = if (elapsed > 0) elapsed else 0L
+        // unchanged rather than counting backwards. The elapsed count is
+        // narrowed to Int here — everything it feeds is an Int (the stored
+        // cycle day and days-until estimate), and a realistic elapsed day
+        // count is nowhere near Int range.
+        val rolled = if (elapsed > 0) elapsed.toInt() else 0
         val day = baseDay + rolled
         // The countdown stops rendering once it would cross zero: the app
         // is the only authority for a fresh estimate.
